Avoid partial list functions in chart code
The old definition of `adjacentPairs` is actually perfectly safe anyway, due to laziness, but still it's easy to rewrite it to avoid using `tail`.

As for `head bottoms`, we could accept only non-empty lists and thread through a proof that `bottoms` remains non-empty. But this would require going through some doubly-non-empty type for expressing that `scanl` increases the length of a list by one, and `adjacentPairs` reduces by one again...