Bump diagrams fork
This causes some trivial changes to existing files due to a fix to transform origins. There are some small changes to the values for the scaling animations for Petri place fillings, which actually fix some subtle off-centering. Plus some big changes for the translation animations for the vertical time bars on charts, which cause no visual difference, as the origin doesn't matter for translations.

This bump also brings us support for rotation animations.